+++ src/test/clojure/clara/test_rules.clj
@@ -17,7 +17,7 @@
             [clara.order-ruleset :as order-rules]
             [schema.test]
             [schema.core :as sc]
-            [clara.tools.testing-utils :as tu])
+            [clara.tools.testing-utils :as tu :refer [assert-ex-data]])
   (:import [clara.rules.testfacts Temperature WindSpeed Cold Hot TemperatureHistory
             ColdAndWindy LousyWeather First Second Third Fourth FlexibleFields]
            [clara.rules.engine
@@ -40,54 +40,6 @@
 
 (defn- has-fact? [token fact]
   (some #{fact} (map first (:matches token))))
-
-(defn ex-data-search [^Exception e edata]
-  (loop [non-matches []
-         e e]
-    (cond
-      ;; Found match.
-      (= edata
-         (select-keys (ex-data e)
-                      (keys edata)))
-      :success
-
-      ;; Keep searching, record any non-matching ex-data.
-      (.getCause e)
-      (recur (if-let [ed (ex-data e)]
-               (conj non-matches ed)
-               non-matches)
-             (.getCause e))
-
-      ;; Can't find a match.
-      :else
-      non-matches)))
-
-(defn get-all-ex-data
-  "Walk a Throwable chain and return a sequence of all data maps
-  from any ExceptionInfo instances in that chain."
-  [e]
-  (let [get-ex-chain (fn get-ex-chain [e]
-                       (if-let [cause (.getCause e)]
-                         (conj (get-ex-chain cause) e)
-                         [e]))]
-
-    (map ex-data
-         (filter (partial instance? clojure.lang.IExceptionInfo)
-                 (get-ex-chain e)))))
-
-(defmacro assert-ex-data [expected-ex-data form]
-  `(try
-     ~form
-     (is false
-         (str "Exception expected to be thrown when evaluating: " \newline
-              '~form))
-     (catch Exception e#
-       (let [res# (ex-data-search e# ~expected-ex-data)]
-         (is (= :success res#)
-             (str "Exception msg found: " \newline
-                  e# \newline
-                  "Non matches found: " \newline
-                  res#))))))
 
 (deftest test-malformed-binding
   ;; Test binding with no value.
@@ -1568,7 +1520,7 @@
                                    (fire-rules)))
 
         listeners-from-trace (fn [e] (mapcat :listeners
-                                             (get-all-ex-data e)))]
+                                             (tu/get-all-ex-data e)))]
     (try
       (run-session-traced)
       (is false "Running the rules in this test should cause an exception.")